logic fixed, added main, added division by tasks
