+++ src/appeal_card.py
@@ -1,7 +1,6 @@
 class AppealCard:
     def __init__(self, number, date, author, email, telephone,
                  city, address, social_status, addressee, appeal_text, category):
-        print("Hello from __init__")
         self.number = number
         self.date = date
         self.author = author
