+++ src/pdf_extractor.py
@@ -1,0 +1,5 @@
+import pdfplumber
+
+def extract_text_from_pdf(pdf_path):
+    with pdfplumber.open(pdf_path) as pdf:
+        return "\n".join(page.extract_text() for page in pdf.pages if page.extract_text())
